move ase conversions to own file
